+++ tests/conftest.py
@@ -34,3 +34,33 @@ def fixture_remotepass_soup() -> BeautifulSoup:
     soup = BeautifulSoup(html, "lxml")
     clean_html(soup)
     return soup
+
+
+@pytest.fixture(scope="session", name="rappipay_in_soup")
+def fixture_rappipay_in_soup() -> BeautifulSoup:
+    html = Path("tests/html_samples/rappipay_bank_transfer_in.html").read_text(
+        encoding="utf-8"
+    )
+    soup = BeautifulSoup(html, "lxml")
+    clean_html(soup)
+    return soup
+
+
+@pytest.fixture(scope="session", name="rappipay_out_soup")
+def fixture_rappipay_out_soup() -> BeautifulSoup:
+    html = Path("tests/html_samples/rappipay_bank_transfer_out.html").read_text(
+        encoding="utf-8"
+    )
+    soup = BeautifulSoup(html, "lxml")
+    clean_html(soup)
+    return soup
+
+
+@pytest.fixture(scope="session", name="rappipay_pse_soup")
+def fixture_rappipay_pse_soup() -> BeautifulSoup:
+    html = Path("tests/html_samples/rappipay_pse_payment.html").read_text(
+        encoding="utf-8"
+    )
+    soup = BeautifulSoup(html, "lxml")
+    clean_html(soup)
+    return soup